third_party/WebKit/LayoutTests/platform/win/custom-elements/spec/state-failed-create-expected.txt

CONSOLE ERROR: line 2: Uncaught Error
This is a testharness.js-based test.
PASS Custom Elements: Custom Element State "Failed" in document parser 
Harness: the test ran to completion.
